DESK-421 disable hosted if remote and not running with elevated permissions
