+++ frontend/src/components/SetupView/index.ts
@@ -1,0 +1,1 @@
+export { SetupView } from './SetupView'
